+++ application/pivot/daanse.pivot.bndrun
@@ -28,7 +28,6 @@
 -runfw: org.apache.felix.framework
 
 -runrequires: \
-	bnd.identity;id='${project.artifactId}',\
 	bnd.identity;id='org.glassfish.hk2.osgi-resource-locator',\
 	bnd.identity;id='com.sun.xml.messaging.saaj.impl',\
 	bnd.identity;id='org.apache.aries.spifly.dynamic.framework.extension',\
@@ -36,14 +35,10 @@
 	bnd.identity;id='org.eclipse.daanse.olap.xmla.bridge',\
 	bnd.identity;id='org.eclipse.daanse.xmla.server.jakarta.saaj',\
 	bnd.identity;id='ch.qos.logback.classic',\
-	bnd.identity;id='org.eclipse.daanse.jdbc.datasource.metatype.h2',\
-	bnd.identity;id='org.eclipse.daanse.jdbc.db.dialect.db.h2',\
-	bnd.identity;id='org.eclipse.daanse.jdbc.db.core',\
 	bnd.identity;id='org.eclipse.daanse.mdx.parser.api',\
 	bnd.identity;id='org.eclipse.daanse.mdx.parser.ccc',\
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.api',\
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.emf',\
-	bnd.identity;id='org.eclipse.daanse.jdbc.db.importer.csv',\
 	bnd.identity;id='org.gecko.emf.osgi.component',\
 	bnd.identity;id='org.eclipse.daanse.sql.guard.api',\
 	bnd.identity;id='org.eclipse.daanse.sql.guard.jsqltranspiler',\
@@ -59,9 +54,7 @@
 	ai.starlake.transpiler;version='[1.0.0,1.0.1)',\
 	ch.qos.logback.classic;version='[1.5.6,1.5.7)',\
 	com.github.ben-manes.caffeine;version='[3.1.8,3.1.9)',\
-	com.h2database;version='[2.3.230,2.3.231)',\
 	com.sun.xml.messaging.saaj.impl;version='[3.0.3,3.0.4)',\
-	de.siegmar.fastcsv;version='[3.1.0,3.1.1)',\
 	jakarta.activation-api;version='[2.1.2,2.1.3)',\
 	jakarta.xml.soap-api;version='[3.0.1,3.0.2)',\
 	json;version='[20250517.0.0,20250517.0.1)',\
@@ -72,18 +65,8 @@
 	org.apache.felix.http.servlet-api;version='[3.0.0,3.0.1)',\
 	org.apache.felix.metatype;version='[1.2.4,1.2.5)',\
 	org.apache.felix.scr;version='[2.2.10,2.2.11)',\
-	org.eclipse.daanse.io.fs.watcher.api;version='[0.0.1,0.0.2)',\
-	org.eclipse.daanse.io.fs.watcher.watchservice;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.jakarta.servlet.soap;version='[0.0.1,0.0.2)',\
-	org.eclipse.daanse.jdbc.datasource.metatype.common;version='[0.0.1,0.0.2)',\
-	org.eclipse.daanse.jdbc.datasource.metatype.h2;version='[0.0.1,0.0.2)',\
-	org.eclipse.daanse.jdbc.db.api;version='[0.0.1,0.0.2)',\
-	org.eclipse.daanse.jdbc.db.core;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.jdbc.db.dialect.api;version='[0.0.1,0.0.2)',\
-	org.eclipse.daanse.jdbc.db.dialect.db.common;version='[0.0.1,0.0.2)',\
-	org.eclipse.daanse.jdbc.db.dialect.db.h2;version='[0.0.1,0.0.2)',\
-	org.eclipse.daanse.jdbc.db.importer.csv;version='[0.0.1,0.0.2)',\
-	org.eclipse.daanse.jdbc.db.record;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.mdx.model.api;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.mdx.model.record;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.mdx.parser.api;version='[0.0.1,0.0.2)',\
@@ -97,7 +80,6 @@
 	org.eclipse.daanse.rolap.mapping.api;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.rolap.mapping.emf;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.rolap.mapping.pojo;version='[0.0.1,0.0.2)',\
-	org.eclipse.daanse.server.application.probebox;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.sql.guard.api;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.sql.guard.jsqltranspiler;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.xmla.api;version='[0.0.1,0.0.2)',\
