+++ application/pivot/pom.xml
@@ -1,0 +1,56 @@
+<?xml version="1.0"?>
+<!--
+/*********************************************************************
+* Copyright (c) 2024 Contributors to the Eclipse Foundation.
+*
+* This program and the accompanying materials are made
+* available under the terms of the Eclipse Public License 2.0
+* which is available at https://www.eclipse.org/legal/epl-2.0/
+*
+* SPDX-License-Identifier: EPL-2.0
+**********************************************************************/
+-->
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.eclipse.daanse</groupId>
+    <artifactId>org.eclipse.daanse.server.application</artifactId>
+    <version>${revision}</version>
+  </parent>
+  <artifactId>org.eclipse.daanse.server.application.pivot</artifactId>
+  
+  
+  
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>biz.aQute.bnd</groupId>
+        <artifactId>bnd-export-maven-plugin</artifactId>
+        <version>${bnd.version}</version>
+        <configuration>
+          <failOnChanges>false</failOnChanges>
+          <bndruns>
+            <bndrun>daanse.pivot.bndrun</bndrun>
+          </bndruns>
+          <exporter>bnd.executablejar</exporter>
+          <includeDependencyManagement>true</includeDependencyManagement>
+          <reportOptional>false</reportOptional>
+          <scopes>
+            <scope>compile</scope>
+            <scope>runtime</scope>
+            <scope>test</scope>
+          </scopes>
+        </configuration>
+        <executions>
+          <execution>
+            <goals>
+              <goal>export</goal>
+            </goals>
+          </execution>
+        </executions>
+      </plugin>
+    </plugins>
+  </build>
+</project>
