+++ application/pom.xml
@@ -23,7 +23,8 @@
   <packaging>pom</packaging>
 
   <modules>
-    <module>probebox</module>
+    <module>probe</module>
+    <module>pivot</module>
   </modules>
 
   <dependencies>
@@ -32,6 +33,13 @@
       <artifactId>org.apache.felix.fileinstall</artifactId>
       <version>3.7.4</version>
       <scope>runtime</scope>
+    </dependency>
+
+    <dependency>
+      <groupId>org.eclipse.daanse</groupId>
+      <artifactId>org.eclipse.daanse.rolap.core</artifactId>
+      <version>0.0.1-SNAPSHOT</version>
+      <scope>compile</scope>
     </dependency>
 
     <dependency>
